requirements.txt
google-genai>=0.3.0
msal>=1.24.0
requests>=2.31.0
pymongo>=4.6.0
python-docx>=0.8.11
pdf2image>=1.16.3
pillow>=10.0.0
flask>=3.0.0
flask-cors>=4.0.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
pyjwt==2.8.0